fix: state of negotiation in E2E.
